separate text saves; fix build without TEXTSAVE
